requirements.txt
# Core dependencies
pydantic
anyio
python-dotenv
requests
aiohttp  # Required for async Azure AD authentication

# Azure Computer Vision for OCR
azure-ai-vision-imageanalysis
azure-identity

# Large Model Planner (install what you need)
# For OpenAI or Azure OpenAI (GPT-4o):
openai
# For Anthropic (Claude):
anthropic
# For Google Gemini:
# google-generativeai

# Testing (for development)
pytest>=9.0.0
pytest-asyncio>=1.3.0
pytest-mock>=3.15.0

# Small Model Workers (optional - install what you need)
# For local inference (Phi-3, Llama):
# transformers
# torch
# accelerate
# For Ollama: just install Ollama and use requests (already included)
